modified:   app/code/community/NoPro/Bluemoon/etc/config.xml modified:   app/code/community/NoPro/Bluemoon/etc/system.xml modified:   app/code/community/NoPro/Bluemoon/sql/nopro_bluemoon_setup/upgrade-1.0.0.2-1.0.1.0.php modified:   package.xml
	app/code/community/NoPro/Bluemoon/sql/nopro_bluemoon_setup/upgrade-1.0.1.0-1.0.1.1.php

+++ app/code/community/NoPro/Bluemoon/sql/nopro_bluemoon_setup/upgrade-1.0.0.2-1.0.1.0.php
@@ -29,7 +29,8 @@ foreach($newfields as $f) {
 
 $dropfields = array('n_feld_halt','nochn_feld');
 foreach($dropfields as $f) {
-  $pdo->dropColumn($table, $f);
+  if($pdo->tableColumnExists($table, $f))
+    $pdo->dropColumn($table, $f);
 }
 
 $this->endSetup();
